Dismissing nav now works
Added the ability to use camel case class names in jsx but dashed ones in
css (like it should be).  This was done by adding a flag to
webpack.config.

As I mentioned in a previous commit, the nav and top bar might have to
move out of layout and into a higher level so that they don't flash when
the page changes.