Simplify build process
I completely deleted the run.js file because I think it was overly
complicated and I had no use for .ejs tempaltes.

In order to run it now run "npm start" and navigate to "localhost:8080".
I did lose the browser automatically launching via browsersync, but that
could be added again.

I followed the react-router as a template for allowing hot module swapping
to work again, but this might be a better long term solution:

https://github.com/gaearon/react-hot-boilerplate