Use global @value breakpoints
I tried for a long time to use postcss-modules-values[1] but couldn't get it
to work.  So I then tried postcss-modules-values-replace[2] and it stilll
didn't work.

Finally, thanks to a comment on an issue I figured out that it was the
fact that I was accidentally including two sets of parenthesis around my
media queries.  I wish I could see the compiled css somewhere to be able
to figure that out.

[1]: https://github.com/css-modules/postcss-modules-value://github.com/css-modules/postcss-modules-values
[2]: https://github.com/princed/postcss-modules-values-replac://github.com/princed/postcss-modules-values-replace