Thumbnail aspect ration preserved
The string calculation I was doing to set the width of each thumbnail
was failing and not actually appearing on the final result.  I fixed
this by removing it from the string.

The Card CSS was also changed in the debuggin process to remove the
global image style, in a more CSS modules style.