Removing courses works correctly
The issue was with the key that I was supplying react.  Because I was
using the index, when I removed things and the index changed for all t
he latter elements, it got confused.  To remedy this, I created a simple
count in the state of scheduler that is incremented every time an
Course, Section, or Time is created.  This way the number is alwasy
unique, but also stays consistent with that item as the array is
rearranged.