Sync scheduler state with textbox
This allows the user to enter their class data in the UI, and then save
the contents of the text box.  Then can then at a later time, come back
to the textbox and paste their saved data in there, which will
automatically populate the UI, allowing them to pick up right where they
left off.

React is yelling at me for doing this because I'm allowing the user to
edit the text box while also progrmatically editing it.  I think however
that in this case this is the best solution.