Fix jumbotron margins
Before when I had the entire layout indented form both sides, I had to
give the jumbotron negative margins so that it would stretch across the
whole page.  After removing the margins from the whole page, I had not
fixed the margins on the jumbotron, creating annoying negative scrolling
when it was not necessary.